rename prismada to prisma-da
